refactor: ApplyPass 관련 패키지 구조 및 경로 수정
