+++ src/main/java/org/ject/support/admin/apply/controller/ApplyPassApiSpec.java
@@ -1,8 +1,8 @@
-package org.ject.support.domain.admin.controller;
+package org.ject.support.admin.apply.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import org.ject.support.domain.admin.dto.ApplyPassRequest;
+import org.ject.support.admin.apply.dto.ApplyPassRequest;
 import org.springframework.web.bind.annotation.RequestBody;
 
 @Tag(name = "Apply Pass", description = "지원자 합격 API")
