+++ src/test/java/org/ject/support/admin/apply/service/ApplyPassServiceTest.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.admin.service;
+package org.ject.support.admin.apply.service;
 
 import org.ject.support.base.UnitTestSupport;
 import org.ject.support.domain.apply.domain.ApplicationForm;
